server/ingest/requirements.txt
requests>=2.31.0
pypdf>=3.17.0
python-dotenv>=1.0.1


